pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "photoshop-mcp-server"
version = "0.1.0"
description = "MCP Server for Photoshop integration using photoshop-python-api"
authors = ["Hal <hallong@tencent.com>"]
readme = "README.md"
license = "MIT"
repository = "https://github.com/loonghao/photoshop-python-api-mcp-server"

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
mcp = "^1.6.0"
photoshop-python-api = "^0.24.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
isort = "^5.12.0"
ruff = "^0.0.284"
pre-commit = "^3.3.3"
pytest-cov = "^4.1.0"
pytest-mock = "^3.11.1"
nox = "^2023.4.22"

[tool.poetry.scripts]
ps-mcp = "photoshop_mcp_server.server:main"

[tool.mcp]
name = "Photoshop"
description = "Control Adobe Photoshop using MCP"
version = "0.1.0"
icon = "https://raw.githubusercontent.com/loonghao/photoshop-python-api-mcp-server/main/assets/photoshop-icon.png"
authors = ["Hal <hal.long@outlook.com>"]
repository = "https://github.com/loonghao/photoshop-python-api-mcp-server"
entrypoint = "photoshop_mcp_server.server:create_server"

[tool.black]
line-length = 120
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 120

[tool.ruff]
line-length = 120
target-version = "py38"
select = ["E", "F", "B", "I"]
ignore = []
